pyproject.toml
[tool.poetry]
name = "nonebot-plugin-qbittorrent-manager"
version = "1.1.3"
description = "qbittorrent管理器"
authors = ["SuperGuGuGu <from89917812@163.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "nonebot_plugin_qbittorrent_manager"}]
repository = "https://github.com/SuperGuGuGu/nonebot-plugin-qbittorrent-manager"
keywords = ["nonebot2", "qbittorrent"]

[tool.poetry.dependencies]
nonebot2 = ">=2.2.0"
python = ">=3.10,<4.0"
httpx = ">=0.23"
pillow = ">=11.0.0"
matplotlib = ">=3.9.0"
pathlib = ">=1.0"
nonebot-plugin-send-anything-anywhere = ">=0.7.0"
nonebot-plugin-localstore = ">=0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
